did required fixes

+++ rampup/accounts/views.py
@@ -9,14 +9,10 @@ from rest_framework.authentication import (
 from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.authtoken.models import Token
-from rest_framework.response import Response
-from rest_framework import mixins
 
 from .models import User
-from restaurant.models import Restaurant, ResFoodItem, OrderedItem, Order
 from .serializers import (
     UserSerializer, RegisterSerializer, LoginSerializer, 
-    RestaurantSerializer, RestaurantItemSerializer, OrderSerializer
 )
 
 class UserViewSet(viewsets.ModelViewSet):
@@ -69,8 +65,6 @@ class LogoutView(APIView):
     authentication_classes = (TokenAuthentication, )
    
     def post(self, request):
-
-
         auth = get_authorization_header(request).split()
         if not auth:
             raise exceptions.AuthenticationFailed('Authenticate First')
@@ -86,54 +80,9 @@ class LogoutView(APIView):
         except UnicodeError:
             msg = ('Invalid token header. Token string should not contain invalid characters.')
             raise exceptions.AuthenticationFailed(msg)
-        token_object = Token.objects.get(key=token)
-        token_object.delete()
-
+        token_object = Token.objects.filter(key=token)
+        if token_object:
+            token_object.delete()
+        else:
+            raise exceptions.AuthenticationFailed("Given Token not associated with the user")
         return Response(status=204)
-
-
-class RestaurantViewSet(viewsets.ModelViewSet):
-    """
-    To handles CRUD operation on resturants
-    """
-    permission_classes = [permissions.IsAuthenticated]
-    queryset = Restaurant.objects.all()
-    serializer_class = RestaurantSerializer  
-
-
-class RestaurantItemViewSet(viewsets.ModelViewSet):
-    """
-    To handle items present in a restaurant
-    """
-    serializer_class = RestaurantItemSerializer
-    
-    def get_permissions(self):
-        self.permission_classes = [permissions.IsAuthenticated]
-        given_id = self.request.query_params.get('id')
-        if given_id is None:
-            raise exceptions.AuthenticationFailed("provide restaurant id with id as a parameter")
-        restaurant_object = Restaurant.objects.get(pk=given_id)
-        if self.request.user not in restaurant_object.owner_ids.all():
-            raise exceptions.AuthenticationFailed("Not an owner")
-        return super().get_permissions()
-        
-    def get_queryset(self):
-        given_id = self.request.query_params.get('id')
-        queryset = ResFoodItem.objects.filter(res_id=given_id)
-        return queryset
-
-
-
-class OrderViewSet(viewsets.ModelViewSet):
-      
-    """
-    To handle the orders
-    """
-    permission_classes = [permissions.IsAuthenticated]
-    
-    queryset = Order.objects.all()
-    serializer_class = OrderSerializer   
-    def list(self, request):
-        queryset = Order.objects.filter(user_id=request.user)
-        serializer = OrderSerializer(queryset, many=True)
-        return Response(serializer.data)
